requirements/base.txt
pytz==2021.1  # https://github.com/stub42/pytz
python-slugify==4.0.1  # https://github.com/un33k/python-slugify
Pillow==8.1.2  # https://github.com/python-pillow/Pillow
argon2-cffi==20.1.0  # https://github.com/hynek/argon2_cffi
redis==3.5.3  # https://github.com/andymccurdy/redis-py
hiredis==1.1.0  # https://github.com/redis/hiredis-py
celery  # pyup: < 5.0,!=4.4.7  # https://github.com/celery/celery
django-celery-beat  # https://github.com/celery/django-celery-beat
flower  # https://github.com/mher/flower
uvicorn[standard]==0.13.4  # https://github.com/encode/uvicorn
drf-yasg
Twisted==20.3.0
django-elasticsearch-dsl
python-logstash
kubernetes
django_celery_results
sqlalchemy
djoser
stripe
dj-stripe
channels
channels_redis
djangochannelsrestframework
docker
influxdb-client
django_otp
qrcode
# Django
# ------------------------------------------------------------------------------
django==3.1.7  # pyup: < 3.2  # https://www.djangoproject.com/
djoser
django-mjml==0.12.0
django-post-office==3.6.0
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==4.1.1  # https://github.com/jazzband/django-model-utils
django-allauth==0.44.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.11.2  # https://github.com/django-crispy-forms/django-crispy-forms
django-redis==4.12.1  # https://github.com/jazzband/django-redis
# Django REST Framework
djangorestframework_simplejwt
djangorestframework==3.12.4  # https://github.com/encode/django-rest-framework
django-cors-headers==3.7.0 # https://github.com/adamchainz/django-cors-headers
sentry-sdk==1.0.0
google-cloud-container==2.4.1
pip~=21.1.2
setuptools~=57.0.0
certifi~=2021.5.30
urllib3~=1.26.5
requests~=2.25.1
chardet~=4.0.0
idna~=2.10
django-anymail[mailgun]==8.2  # https://github.com/anymail/django-anymail

